mensajes totalmente terminado
