+++ routes/web.php
@@ -29,7 +29,7 @@ Route::prefix('publicos')->middleware('auth')->group(function(){
     Route::get('/ProductosNegativos','publico\PublicoController@filtarProductosNegativos')->name('ProductosNegativos');
     Route::get('/ProductosNegativos2','publico\PublicoController@listarFiltrados')->name('filtro2');
     Route::get('/Informacion','publico\PublicoController@informacion')->name('informacion');
-    
+    Route::post('/updatemensaje', 'publico\PublicoController@updatemensaje')->name('updatemensaje');
 
 
 
@@ -59,7 +59,6 @@ Route::prefix('admin')->namespace('Admin')->middleware('auth','SuperAdmin')->gro
     Route::get('/','AdminController@index')->name('inicioAdmin');
     Route::get('/ListaUsuarios','EditarUserController@index')->name('ListarUser');
     Route::post('/update', 'EditarUserController@update')->name('update');
-    Route::post('/updatemensaje', 'AdminController@updatemensaje')->name('updatemensaje');
     Route::get('/CuadroMando', 'AdminController@CuadroDeMando')->name('cuadroMando');
     Route::get('/ProductosPorMarca','AdminController@ProductosPorMarca')->name('ProductosPorMarca');
     Route::get('/Ordenesdecompra','AdminController@ordenesdecompra')->name('ordenesdecompra');
